Werror Conversion Flag Enabled (#449)
Enables -Werror=conversion for firmware and fixes every resulting
warning across the codebase, except in phal_F4, sdio, and
third-party/generated code, which are explicitly exempted.

- Added -Werror=conversion (firmware/cmake/toolchain.cmake) 

- Fixed all conversion warnings found across common/ and every node
under source/

- Left phal_F4 and sdio untouched beyond what was required to compile 

- Converted several previously-INTERFACE libraries to STATIC (phal_F4,
phal_G4, sdio, external/fatfs, external/w5500) so each can carry its own
PRIVATE -Wno-conversion compile options

- Suppressed conversion warnings per-library for code not worth cleaning
up: phal_F4, sdio, fatfs, w5500 (third-party/vendor), plus
torque_vector/vcu/* (Matlab/Simulink-generated, not hand-maintained).

---------

+++ firmware/common/heartbeat/heartbeat.h
@@ -28,11 +28,11 @@ typedef enum : uint8_t {
 
 typedef struct {
     GPIO_TypeDef *heartbeat_port;
-    uint32_t heartbeat_pin;
+    uint8_t heartbeat_pin;
     GPIO_TypeDef *connection_port;
-    uint32_t connection_pin;
+    uint8_t connection_pin;
     GPIO_TypeDef *error_port;
-    uint32_t error_pin;
+    uint8_t error_pin;
     void (*preflight_callback)(void);
     heartbeat_state_t state;
 } status_leds_t;
