+++ firmware/source/a_box/adbms/pec.h
@@ -12,8 +12,9 @@
 
 #include <stdbool.h>
 #include <stdint.h>
+#include <stddef.h>
 
 uint16_t adbms_pec_get_pec15(uint8_t len, const uint8_t *data);
-uint16_t adbms_pec_get_pec10(bool bIsRxCmd, uint8_t nLength, const uint8_t *pDataBuf);
+uint16_t adbms_pec_get_pec10(bool bIsRxCmd, size_t nLength, const uint8_t *pDataBuf);
 
 #endif // PEC_H
